Add optional fine grained logging to the build pipeline
The log gets written to a `dist-newstyle/cache/timing-log.txt` and
format is "Start time, action, duration in seconds". The output is
not sorted on start time because the log tuple is only written at
when the action is complete, so a fast action that starts after
a slow action might be written before the fast action.

Actions other than `build` could probably also be instrumented in
a similar way, but that is not part of this PR.

Example log line:

  2024/06/18 06:29:30 UTC: Downloading colour-2.3.6 took 0.815885883s

An example command to enable the timming log (disabled by default):

  cabal --enable-log-timing build all