Introduce AbsolutePath to symbolic path abstraction
There are a few places where paths are known to be absolute. This
enforces that in the type system by introducing a simple wrapper to
`Distribution.Utils.Path`.

```
newtype AbsolutePath (to :: FileOrDir) = AbsolutePath (forall from . SymbolicPath from to)
```

The nice thing about this abstraction is when when a path is unwrapped,
due to the universally quantified `from` type, the resulting
`SymbolicPath` can be used with any API which expects a path to point
`from` a specific directory.